# whatsapp-api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp Web API",
    "description": "API para integração com WhatsApp Web utilizando whatsapp-web.js. Permite conexão de múltiplos usuários, envio de mensagens e gerenciamento de sessões.",
    "version": "1.0.0",
    "contact": {
      "name": "WhatsApp Web API",
      "description": "Microserviço para integração com WhatsApp"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://api.whatsapp-web.com",
      "description": "Servidor de produção"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
    "/api/whatsapp/connect": {
      "post": {
        "tags": ["Conexões"],
        "summary": "Criar nova conexão WhatsApp",
        "description": "Inicia uma nova sessão do WhatsApp Web para um usuário específico. Retorna informações sobre a sessão criada.",
        "operationId": "createConnection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sessionId", "data"],
                "properties": {
                  "sessionId": {
                    "type": "string",
                    "description": "Identificador único da sessão do usuário",
                    "example": "user123"
                  },
                  "data": {
                    "type": "object",
                    "required": ["notifications_url", "messages_url"],
                    "description": "Dados de configuração da sessão (extensível para campos futuros)",
                    "properties": {
                      "notifications_url": {
                        "type": "string",
                        "description": "URL para receber eventos de ciclo de vida (qr_code, authenticated, ready, disconnected, error, loading_screen)",
                        "example": "https://meu-app.com/webhooks/notifications"
                      },
                      "messages_url": {
                        "type": "string",
                        "description": "URL para receber mensagens recebidas (event: message)",
                        "example": "https://meu-app.com/webhooks/messages"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Conexão criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "sessionId": {
                      "type": "string",
                      "example": "user123"
                    },
                    "message": {
                      "type": "string",
                      "example": "Conexão criada com sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "sessionId é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/whatsapp/qrcode/{sessionId}": {
      "get": {
        "tags": ["QR Code"],
        "summary": "Obter QR Code de uma sessão",
        "description": "Retorna o QR Code necessário para autenticar a sessão do WhatsApp Web.",
        "operationId": "getQRCode",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "Identificador da sessão",
            "schema": {
              "type": "string"
            },
            "example": "user123"
          }
        ],
        "responses": {
          "200": {
            "description": "QR Code gerado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "sessionId": {
                      "type": "string",
                      "example": "user123"
                    },
                    "qr": {
                      "type": "string",
                      "description": "String do QR Code gerado pelo WhatsApp",
                      "example": "2@abc123..."
                    },
                    "qrImage": {
                      "type": "string",
                      "description": "QR Code em formato base64 (data URL)",
                      "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-15T10:30:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "sessionId não fornecido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "sessionId é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Sessão não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Sessão não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/whatsapp/status/{sessionId}": {
      "get": {
        "tags": ["Status"],
        "summary": "Verificar status da conexão",
        "description": "Retorna o status atual da conexão WhatsApp para uma sessão específica.",
        "operationId": "getConnectionStatus",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "Identificador da sessão",
            "schema": {
              "type": "string"
            },
            "example": "user123"
          }
        ],
        "responses": {
          "200": {
            "description": "Status da conexão",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "sessionId": {
                      "type": "string",
                      "example": "user123"
                    },
                    "status": {
                      "$ref": "#/components/schemas/ConnectionStatus"
                    },
                    "info": {
                      "type": "object",
                      "nullable": true,
                      "description": "Informações do cliente WhatsApp quando conectado"
                    },
                    "isBusiness": {
                      "type": "boolean",
                      "description": "True se a sessão conectada for WhatsApp Business (Conn.platform smba/smbi)",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "sessionId não fornecido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "sessionId é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Sessão não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Sessão não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/whatsapp/send/{sessionId}": {
      "post": {
        "tags": ["Mensagens"],
        "summary": "Enviar mensagem",
        "description": "Envia texto e/ou imagem para um destinatário usando uma sessão autenticada. Para imagem, envie `media` com `mimetype` (image/*), `data` (base64) e opcionalmente `filename`. `message` vira legenda quando há mídia.",
        "operationId": "sendMessage",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "Identificador da sessão",
            "schema": {
              "type": "string"
            },
            "example": "user123"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["to"],
                "properties": {
                  "to": {
                    "type": "string",
                    "description": "JID do destinatário no WhatsApp (ex.: 5511999999999@c.us). Preferir o `jid` do webhook de message. Passado diretamente para client.getChatById().",
                    "example": "5511999999999@c.us"
                  },
                  "message": {
                    "type": "string",
                    "description": "Texto da mensagem, ou legenda quando `media` está presente",
                    "example": "Olá! Esta é uma mensagem de teste."
                  },
                  "media": {
                    "type": "object",
                    "description": "Imagem em base64 (apenas image/*)",
                    "properties": {
                      "mimetype": {
                        "type": "string",
                        "example": "image/jpeg"
                      },
                      "data": {
                        "type": "string",
                        "description": "Conteúdo base64 (com ou sem prefixo data:)"
                      },
                      "filename": {
                        "type": "string",
                        "example": "foto.jpg"
                      }
                    },
                    "required": ["mimetype", "data"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mensagem enviada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "messageId": {
                      "type": "string",
                      "description": "ID da mensagem enviada",
                      "example": "3EB0C767D82B0A78"
                    },
                    "to": {
                      "type": "string",
                      "description": "JID do destinatário usado no envio",
                      "example": "5511999999999@c.us"
                    },
                    "message": {
                      "type": "string",
                      "example": "Olá! Esta é uma mensagem de teste."
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-15T10:30:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou sessão não autenticada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "sessionId, to e message são obrigatórios"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/whatsapp/disconnect/{sessionId}": {
      "delete": {
        "tags": ["Conexões"],
        "summary": "Desconectar cliente",
        "description": "Desconecta e remove uma sessão específica do WhatsApp Web.",
        "operationId": "disconnectClient",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "Identificador da sessão a ser desconectada",
            "schema": {
              "type": "string"
            },
            "example": "user123"
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente desconectado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "sessionId": {
                      "type": "string",
                      "example": "user123"
                    },
                    "message": {
                      "type": "string",
                      "example": "Cliente desconectado com sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "sessionId não fornecido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "sessionId é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Sessão não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Sessão não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/whatsapp/connections": {
      "get": {
        "tags": ["Conexões"],
        "summary": "Listar todas as conexões",
        "description": "Retorna uma lista de todas as sessões ativas e suas informações.",
        "operationId": "listConnections",
        "responses": {
          "200": {
            "description": "Lista de conexões",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "connections": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "sessionId": {
                            "type": "string",
                            "example": "user123"
                          },
                          "status": {
                            "$ref": "#/components/schemas/ConnectionStatus"
                          },
                          "info": {
                            "type": "object",
                            "nullable": true,
                            "description": "Informações do cliente WhatsApp quando conectado"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Número total de conexões",
                      "example": 3
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionStatus": {
        "type": "string",
        "enum": ["connecting", "connected", "disconnected"],
        "description": "Status em memória do cliente WhatsApp"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic Auth (BASIC_AUTH_USERNAME / BASIC_AUTH_PASSWORD)"
      }
    }
  },
...
}